+++ src/FlowtideDotNet.Core/Operators/Join/NestedLoopJoin/BlockNestedJoinOperator.cs
@@ -10,62 +10,68 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using FlowtideDotNet.Base.Vertices.MultipleInput;
-using FlowtideDotNet.Storage.StateManager;
-using System.Threading.Tasks.Dataflow;
-using FlowtideDotNet.Storage.Tree;
-using FlowtideDotNet.Core.Operators.Join.MergeJoin;
-using FlowtideDotNet.Core.Storage;
 using FlowtideDotNet.Base;
-using FlowtideDotNet.Substrait.Relations;
-using FlexBuffers;
-using System.Buffers;
-using Microsoft.Extensions.Logging;
-using System.Diagnostics;
-using FlowtideDotNet.Core.Compute.Internal;
-using FlowtideDotNet.Core.Compute;
 using FlowtideDotNet.Base.Metrics;
+using FlowtideDotNet.Base.Vertices.MultipleInput;
+using FlowtideDotNet.Core.ColumnStore;
+using FlowtideDotNet.Core.ColumnStore.DataValues;
+using FlowtideDotNet.Core.ColumnStore.TreeStorage;
+using FlowtideDotNet.Core.Compute;
+using FlowtideDotNet.Core.Compute.Columnar;
+using FlowtideDotNet.Core.Operators.Join.MergeJoin;
 using FlowtideDotNet.Core.Utils;
-using FlowtideDotNet.Storage.Serializers;
+using FlowtideDotNet.Storage.DataStructures;
+using FlowtideDotNet.Storage.StateManager;
+using FlowtideDotNet.Storage.Tree;
+using FlowtideDotNet.Substrait.CustomProtobuf;
+using FlowtideDotNet.Substrait.Relations;
+using SqlParser.Tokens;
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Threading.Tasks.Dataflow;
 
 namespace FlowtideDotNet.Core.Operators.Join.NestedLoopJoin
 {
-    /// <summary>
-    /// Implemented with 4 trees, 2 persistent, and 2 temporary.
-    /// Data is stored in the temporary trees until a watermark is recieved and it processes the data.
-    /// </summary>
     internal class BlockNestedJoinOperator : MultipleInputVertex<StreamEventBatch, JoinState>
     {
-        // Persisted trees
-        protected IBPlusTree<RowEvent, JoinStorageValue, ListKeyContainer<RowEvent>, ListValueContainer<JoinStorageValue>>? _leftTree;
-        protected IBPlusTree<RowEvent, JoinStorageValue, ListKeyContainer<RowEvent>, ListValueContainer<JoinStorageValue>>? _rightTree;
-
-        // Temporary trees
-        protected IBPlusTree<RowEvent, JoinStorageValue, ListKeyContainer<RowEvent>, ListValueContainer<JoinStorageValue>>? _leftTemporary;
-        protected IBPlusTree<RowEvent, JoinStorageValue, ListKeyContainer<RowEvent>, ListValueContainer<JoinStorageValue>>? _rightTemporary;
-        protected readonly Func<RowEvent, RowEvent, bool> _condition;
-        private readonly JoinRelation joinRelation;
-        private readonly IRowData _rightNullData;
-
+        private readonly JoinRelation _joinRelation;
+        protected readonly Func<EventBatchData, int, EventBatchData, int, bool> _condition;
         private ICounter<long>? _eventsProcessed;
+        private ICounter<long>? _eventsOutCounter;
+        protected IBPlusTree<ColumnRowReference, JoinWeights, ColumnKeyStorageContainer, JoinWeightsValueContainer>? _leftTree;
+        protected IBPlusTree<ColumnRowReference, JoinWeights, ColumnKeyStorageContainer, JoinWeightsValueContainer>? _rightTree;
 
-        public BlockNestedJoinOperator(JoinRelation joinRelation, FunctionsRegister functionsRegister, ExecutionDataflowBlockOptions executionDataflowBlockOptions) : base(2, executionDataflowBlockOptions)
+        protected IBPlusTree<ColumnRowReference, JoinWeights, ColumnKeyStorageContainer, JoinWeightsValueContainer>? _leftTemporary;
+        protected IBPlusTree<ColumnRowReference, JoinWeights, ColumnKeyStorageContainer, JoinWeightsValueContainer>? _rightTemporary;
+
+        private List<int> _leftOutputColumns;
+        private List<int> _rightOutputColumns;
+
+        private List<int> _leftOutputIndices;
+        private List<int> _rightOutputIndices;
+
+
+        public BlockNestedJoinOperator(JoinRelation joinRelation, IFunctionsRegister functionsRegister, ExecutionDataflowBlockOptions executionDataflowBlockOptions) : base(2, executionDataflowBlockOptions)
         {
+            this._joinRelation = joinRelation;
+
             if (joinRelation.Expression == null)
             {
                 throw new InvalidOperationException("Join relation must have an expression");
             }
-            _condition = BooleanCompiler.Compile<RowEvent>(joinRelation.Expression, functionsRegister, joinRelation.Left.OutputLength);
-            this.joinRelation = joinRelation;
 
-            _rightNullData = RowEvent.Create(0, 0, v =>
-            {
-                for (int i = 0; i < joinRelation.Right.OutputLength; i++)
-                {
-                    v.AddNull();
-                }
-            }).RowData;
+            _condition = ColumnBooleanCompiler.CompileTwoInputs(joinRelation.Expression, functionsRegister, joinRelation.Left.OutputLength);
+
+            (_leftOutputColumns, _leftOutputIndices) = GetOutputColumns(joinRelation, 0, joinRelation.Left.OutputLength);
+            (_rightOutputColumns, _rightOutputIndices) = GetOutputColumns(joinRelation, joinRelation.Left.OutputLength, joinRelation.Right.OutputLength);
+
+            GetOutputColumns(joinRelation, 0, joinRelation.Left.OutputLength);
         }
+
 
         public override string DisplayName => "Nested Loop Join";
 
@@ -90,75 +96,65 @@ namespace FlowtideDotNet.Core.Operators.Join.NestedLoopJoin
             return new JoinState();
         }
 
-        protected RowEvent OnConditionSuccess(RowEvent left, RowEvent right, in int weight)
-        {
-            return new RowEvent(weight, 0, ArrayRowData.Create(left.RowData, right.RowData, joinRelation.Emit));
-        }
-
-        protected RowEvent CreateLeftWithNullRightEvent(int weight, RowEvent e)
-        {
-            return new RowEvent(weight, 0, ArrayRowData.Create(e.RowData, _rightNullData, joinRelation.Emit));
-        }
-
         protected override async IAsyncEnumerable<StreamEventBatch> OnWatermark(Watermark watermark)
         {
             Debug.Assert(_leftTree != null, nameof(_leftTree));
             Debug.Assert(_rightTree != null, nameof(_rightTree));
             Debug.Assert(_leftTemporary != null, nameof(_leftTemporary));
             Debug.Assert(_rightTemporary != null, nameof(_rightTemporary));
+            Debug.Assert(_eventsOutCounter != null, nameof(_eventsOutCounter));
 
-            var rightTempIterator = _rightTemporary.CreateIterator();
+            PrimitiveList<int> weights = new PrimitiveList<int>(MemoryAllocator);
+            PrimitiveList<uint> iterations = new PrimitiveList<uint>(MemoryAllocator);
+            IColumn[] columns = new IColumn[_joinRelation.OutputLength];
 
-            var leftPersistentIterator = _leftTree.CreateIterator();
+            for (int i = 0; i < _joinRelation.OutputLength; i++)
+            {
+                columns[i] = Column.Create(MemoryAllocator);
+            }
+
+            using var rightTempIterator = _rightTemporary.CreateIterator();
+
+            using var leftPersistentIterator = _leftTree.CreateIterator();
             await leftPersistentIterator.SeekFirst();
 
-            List<RowEvent> output = new List<RowEvent>();
-
-            // Do the block nested loop join for the values on the right
             await foreach (var leftPage in leftPersistentIterator)
             {
                 bool leftPageModified = false;
+                var leftPageValues = leftPage.Values;
+                var leftColumns = leftPage.Keys._data.Columns;
 
                 // Must always seek the first page in the iterator
                 await rightTempIterator.SeekFirst();
                 await foreach (var rightTmpPage in rightTempIterator)
                 {
                     bool rightModified = false;
-                    foreach(var leftPageKv in leftPage)
+                    var rightPageValues = rightTmpPage.Values;
+                    var rightColumns = rightTmpPage.Keys._data.Columns;
+
+                    for (int leftIndex = 0; leftIndex < leftPage.Keys.Count; leftIndex++)
                     {
-                        foreach(var rightTmpPageKv in rightTmpPage)
+                        for (int rightIndex = 0; rightIndex < rightTmpPage.Keys.Count; rightIndex++)
                         {
-                            if (_condition(leftPageKv.Key, rightTmpPageKv.Key))
+                            if (_condition(leftPage.Keys._data, leftIndex, rightTmpPage.Keys._data, rightIndex))
                             {
-                                var outputWeight = leftPageKv.Value.Weight * rightTmpPageKv.Value.Weight;
-                                output.Add(OnConditionSuccess(leftPageKv.Key, rightTmpPageKv.Key, outputWeight));
-
-                                // Check if it previously was left values with null right
-                                var previousJoinWeight = leftPageKv.Value.JoinWeight;
-                                rightTmpPageKv.Value.JoinWeight += outputWeight;
-                                leftPageKv.Value.JoinWeight += outputWeight;
-
-                                if (joinRelation.Type == JoinType.Left)
-                                {
-                                    // Check if it was previously a left values null right, then output a negation
-                                    if (previousJoinWeight == 0 && leftPageKv.Value.JoinWeight > 0)
-                                    {
-                                        // Output a negation
-                                        output.Add(CreateLeftWithNullRightEvent(-leftPageKv.Value.Weight, leftPageKv.Key));
-                                    }
-                                    // Went back to be a join with left values and null right
-                                    if (previousJoinWeight > 0 && leftPageKv.Value.JoinWeight == 0)
-                                    {
-                                        output.Add(CreateLeftWithNullRightEvent(leftPageKv.Value.Weight, leftPageKv.Key));
-                                    }
-                                }
-                                
-
-                                rightModified = true;
-                                leftPageModified = true;
+                                HandleLeftJoin(
+                                    in leftIndex,
+                                    in rightIndex,
+                                    in leftPageValues,
+                                    in rightPageValues,
+                                    ref leftPageModified,
+                                    ref rightModified,
+                                    in weights,
+                                    in iterations,
+                                    in columns,
+                                    in leftColumns,
+                                    in rightColumns
+                                    );
                             }
                         }
                     }
+
                     if (rightModified)
                     {
                         await rightTmpPage.SavePage(false);
@@ -174,15 +170,17 @@ namespace FlowtideDotNet.Core.Operators.Join.NestedLoopJoin
             // Insert all right temporary values into the right persitent tree
             await foreach (var rightTmpPage in rightTempIterator)
             {
-                foreach(var kv in rightTmpPage)
+                for (int rightIndex = 0; rightIndex < rightTmpPage.Keys.Count; rightIndex++)
                 {
-                    await _rightTree.RMW(kv.Key, kv.Value, (input, current, exist) =>
+                    var rowRef = new ColumnRowReference() { referenceBatch = rightTmpPage.Keys._data, RowIndex = rightIndex };
+                    var value = rightTmpPage.Values.Get(rightIndex);
+                    var op = await _rightTree.RMWNoResult(rowRef, value, (input, current, exist) =>
                     {
                         if (exist)
                         {
-                            current!.Weight += input!.Weight;
-                            current!.JoinWeight += input!.JoinWeight;
-                            if (current.Weight == 0)
+                            current!.weight += input!.weight;
+                            current!.joinWeight += input!.joinWeight;
+                            if (current.weight == 0)
                             {
                                 return (default, GenericWriteOperation.Delete);
                             }
@@ -191,46 +189,73 @@ namespace FlowtideDotNet.Core.Operators.Join.NestedLoopJoin
                         return (input, GenericWriteOperation.Upsert);
                     });
 
-                    // TODO: On right joins or full outer joins, should check if join weight is 0, if so, output left side null with right side values.
+                    if (value.joinWeight == 0 && (_joinRelation.Type == JoinType.Right || _joinRelation.Type == JoinType.Outer))
+                    {
+                        // Output null values here
+                        weights.Add(value.weight);
+                        iterations.Add(0);
+                        for (int i = 0; i < _rightOutputColumns.Count; i++)
+                        {
+                            var outputColumn = columns[_rightOutputIndices[i]];
+                            outputColumn.InsertRangeFrom(outputColumn.Count, rightTmpPage.Keys._data.Columns[_rightOutputColumns[i]], rightIndex, 1);
+                        }
+                        for (int i = 0; i < _leftOutputIndices.Count; i++)
+                        {
+                            columns[_leftOutputIndices[i]].Add(NullValue.Instance);
+                        }
+                    }
                 }
             }
             await _rightTemporary.Clear();
 
-            var leftTempIterator = _leftTemporary.CreateIterator();
-            
-            var rightPersistentIterator = _rightTree.CreateIterator();
+            using var leftTempIterator = _leftTemporary.CreateIterator();
+
+            using var rightPersistentIterator = _rightTree.CreateIterator();
             await rightPersistentIterator.SeekFirst();
 
-            // Go through left side temporary values against right side persistent values
-            await foreach(var rightPage in rightPersistentIterator)
+            await foreach (var rightPage in rightPersistentIterator)
             {
                 bool rightModified = false;
+                var rightPageValues = rightPage.Values;
+                var rightColumns = rightPage.Keys._data.Columns;
 
-                // Must always seek the first page in the inner iterator
                 await leftTempIterator.SeekFirst();
-                await foreach(var leftTempPage in leftTempIterator)
+
+                await foreach (var leftTempPage in leftTempIterator)
                 {
                     bool leftPageModified = false;
-                    foreach (var rightPageKv in rightPage)
+                    var leftPageValues = leftTempPage.Values;
+                    var leftColumns = leftTempPage.Keys._data.Columns;
+
+                    for (int leftIndex = 0; leftIndex < leftTempPage.Keys.Count; leftIndex++)
                     {
-                        foreach(var leftTempPageKv in leftTempPage)
+                        for (int rightIndex = 0; rightIndex < rightPage.Keys.Count; rightIndex++)
                         {
-                            if (_condition(leftTempPageKv.Key, rightPageKv.Key))
+                            if (_condition(leftTempPage.Keys._data, leftIndex, rightPage.Keys._data, rightIndex))
                             {
-                                var outputWeight = leftTempPageKv.Value.Weight * rightPageKv.Value.Weight;
-                                output.Add(OnConditionSuccess(leftTempPageKv.Key, rightPageKv.Key, outputWeight));
-                                rightPageKv.Value.JoinWeight += outputWeight;
-                                leftTempPageKv.Value.JoinWeight += outputWeight;
-                                rightModified = true;
-                                leftPageModified = true;
+                                HandleLRightsideJoin(
+                                    in leftIndex,
+                                    in rightIndex,
+                                    in leftPageValues,
+                                    in rightPageValues,
+                                    ref leftPageModified,
+                                    ref rightModified,
+                                    in weights,
+                                    in iterations,
+                                    in columns,
+                                    in leftColumns,
+                                    in rightColumns
+                                    );
                             }
                         }
                     }
+
                     if (leftPageModified)
                     {
                         await leftTempPage.SavePage(false);
                     }
                 }
+
                 if (rightModified)
                 {
                     await rightPage.SavePage(false);
@@ -238,17 +263,21 @@ namespace FlowtideDotNet.Core.Operators.Join.NestedLoopJoin
             }
 
             await leftTempIterator.SeekFirst();
-            await foreach(var leftTempPage in leftTempIterator)
+
+            // Insert all left temporary values into the left persitent tree
+            await foreach (var leftTmpPage in leftTempIterator)
             {
-                foreach(var kv in leftTempPage)
+                for (int leftIndex = 0; leftIndex < leftTmpPage.Keys.Count; leftIndex++)
                 {
-                    var (op, val) = await _leftTree.RMW(kv.Key, kv.Value, (input, current, exist) =>
+                    var rowRef = new ColumnRowReference() { referenceBatch = leftTmpPage.Keys._data, RowIndex = leftIndex };
+                    var value = leftTmpPage.Values.Get(leftIndex);
+                    var op = await _leftTree.RMWNoResult(rowRef, value, (input, current, exist) =>
                     {
                         if (exist)
                         {
-                            current!.Weight += input!.Weight;
-                            current!.JoinWeight += input!.JoinWeight;
-                            if (current.Weight == 0)
+                            current!.weight += input!.weight;
+                            current!.joinWeight += input!.joinWeight;
+                            if (current.weight == 0)
                             {
                                 return (default, GenericWriteOperation.Delete);
                             }
@@ -257,21 +286,213 @@ namespace FlowtideDotNet.Core.Operators.Join.NestedLoopJoin
                         return (input, GenericWriteOperation.Upsert);
                     });
 
-                    if (op == GenericWriteOperation.Upsert && val!.JoinWeight == 0 && joinRelation.Type == JoinType.Left)
+                    if (value.joinWeight == 0 && (_joinRelation.Type == JoinType.Left || _joinRelation.Type == JoinType.Outer))
                     {
                         // Output null values here
-                        output.Add(CreateLeftWithNullRightEvent(kv.Value.Weight, kv.Key));
-                    }
-                    else
-                    {
-
+                        weights.Add(value.weight);
+                        iterations.Add(0);
+                        for (int i = 0; i < _leftOutputColumns.Count; i++)
+                        {
+                            var outputColumn = columns[_leftOutputIndices[i]];
+                            outputColumn.InsertRangeFrom(outputColumn.Count, leftTmpPage.Keys._data.Columns[_leftOutputColumns[i]], leftIndex, 1);
+                        }
+                        for (int i = 0; i < _rightOutputIndices.Count; i++)
+                        {
+                            columns[_rightOutputIndices[i]].Add(NullValue.Instance);
+                        }
                     }
                 }
             }
 
             await _leftTemporary.Clear();
 
-            yield return new StreamEventBatch(output, joinRelation.OutputLength);
+            _eventsOutCounter.Add(weights.Count);
+            yield return new StreamEventBatch(new EventBatchWeighted(weights, iterations, new EventBatchData(columns)));
+        }
+
+        private static (List<int> incomingIndices, List<int> outgoingIndex) GetOutputColumns(JoinRelation joinRelation, int relative, int maxSize)
+        {
+            List<int> columns = new List<int>();
+            List<int> outgoingIndices = new List<int>();
+            if (joinRelation.EmitSet)
+            {
+                for (int i = 0; i < joinRelation.Emit.Count; i++)
+                {
+                    var index = joinRelation.Emit[i];
+                    if (index >= relative)
+                    {
+                        index = index - relative;
+                        if (index < maxSize)
+                        {
+                            columns.Add(index);
+                            outgoingIndices.Add(i);
+                        }
+                    }
+
+                }
+            }
+            else
+            {
+                for (int i = 0; i < joinRelation.OutputLength - relative; i++)
+                {
+                    if (i < maxSize)
+                    {
+                        columns.Add(i);
+                        outgoingIndices.Add(i + relative);
+                    }
+                }
+            }
+            return (columns, outgoingIndices);
+        }
+
+        private void HandleLeftJoin(
+            ref readonly int leftIndex,
+            ref readonly int rightIndex,
+            ref readonly JoinWeightsValueContainer leftValues,
+            ref readonly JoinWeightsValueContainer rightValues,
+            ref bool leftModified,
+            ref bool rightModified,
+            ref readonly PrimitiveList<int> weights,
+            ref readonly PrimitiveList<uint> iterations,
+            ref readonly IColumn[] columns,
+            ref readonly IReadOnlyList<IColumn> leftColumns,
+            ref readonly IReadOnlyList<IColumn> rightColumns)
+        {
+            ref var leftWeights = ref leftValues.GetRef(leftIndex);
+            ref var rightWeights = ref rightValues.GetRef(rightIndex);
+            var outputWeight = leftWeights.weight * rightValues.Get(rightIndex).weight;
+
+            // Add output here
+            weights.Add(outputWeight);
+            iterations.Add(0);
+            for (int i = 0; i < _leftOutputColumns.Count; i++)
+            {
+                var outputColumn = columns[_leftOutputIndices[i]];
+                outputColumn.InsertRangeFrom(outputColumn.Count, leftColumns[_leftOutputColumns[i]], leftIndex, 1); 
+            }
+            for (int i = 0; i < _rightOutputColumns.Count; i++)
+            {
+                var outputColumn = columns[_rightOutputIndices[i]];
+                outputColumn.InsertRangeFrom(outputColumn.Count, rightColumns[_rightOutputColumns[i]], rightIndex, 1);
+            }
+
+            // Check if it previously was left values with null right
+            var previousJoinWeight = leftWeights.joinWeight;
+            rightWeights.joinWeight += outputWeight;
+            leftWeights.joinWeight += outputWeight;
+
+            if (_joinRelation.Type == JoinType.Left || _joinRelation.Type == JoinType.Outer)
+            {
+                // Check if it was previously a left values null right, then output a negation
+                if (previousJoinWeight == 0 && leftWeights.joinWeight > 0)
+                {
+                    // Output a negation
+                    weights.Add(-leftWeights.weight);
+                    iterations.Add(0);
+                    for (int i = 0; i < _leftOutputColumns.Count; i++)
+                    {
+                        var outputColumn = columns[_leftOutputIndices[i]];
+                        outputColumn.InsertRangeFrom(outputColumn.Count, leftColumns[_leftOutputColumns[i]], leftIndex, 1);
+                    }
+                    for (int i = 0; i < _rightOutputIndices.Count; i++)
+                    {
+                        columns[_rightOutputIndices[i]].Add(NullValue.Instance);
+                    }
+                }
+                // Went back to be a join with left values and null right
+                if (previousJoinWeight > 0 && leftWeights.joinWeight == 0)
+                {
+                    weights.Add(leftWeights.weight);
+                    iterations.Add(0);
+                    for (int i = 0; i < _leftOutputColumns.Count; i++)
+                    {
+                        var outputColumn = columns[_leftOutputIndices[i]];
+                        outputColumn.InsertRangeFrom(outputColumn.Count, leftColumns[_leftOutputColumns[i]], leftIndex, 1);
+                    }
+                    for (int i = 0; i < _rightOutputIndices.Count; i++)
+                    {
+                        columns[_rightOutputIndices[i]].Add(NullValue.Instance);
+                    }
+                }
+            }
+
+            rightModified = true;
+            leftModified = true;
+        }
+
+        private void HandleLRightsideJoin(
+            ref readonly int leftIndex,
+            ref readonly int rightIndex,
+            ref readonly JoinWeightsValueContainer leftValues,
+            ref readonly JoinWeightsValueContainer rightValues,
+            ref bool leftModified,
+            ref bool rightModified,
+            ref readonly PrimitiveList<int> weights,
+            ref readonly PrimitiveList<uint> iterations,
+            ref readonly IColumn[] columns,
+            ref readonly IReadOnlyList<IColumn> leftColumns,
+            ref readonly IReadOnlyList<IColumn> rightColumns)
+        {
+            ref var leftWeights = ref leftValues.GetRef(leftIndex);
+            ref var rightWeights = ref rightValues.GetRef(rightIndex);
+            var outputWeight = leftWeights.weight * rightValues.Get(rightIndex).weight;
+
+            // Add output here
+            weights.Add(outputWeight);
+            iterations.Add(0);
+            for (int i = 0; i < _leftOutputColumns.Count; i++)
+            {
+                var outputColumn = columns[_leftOutputIndices[i]];
+                outputColumn.InsertRangeFrom(outputColumn.Count, leftColumns[_leftOutputColumns[i]], leftIndex, 1);
+            }
+            for (int i = 0; i < _rightOutputColumns.Count; i++)
+            {
+                var outputColumn = columns[_rightOutputIndices[i]];
+                outputColumn.InsertRangeFrom(outputColumn.Count, rightColumns[_rightOutputColumns[i]], rightIndex, 1);
+            }
+
+            // Check if it previously was left values with null right
+            var previousJoinWeight = rightWeights.joinWeight;
+            rightWeights.joinWeight += outputWeight;
+            leftWeights.joinWeight += outputWeight;
+
+            if (_joinRelation.Type == JoinType.Right || _joinRelation.Type == JoinType.Outer)
+            {
+                // Check if it was previously a left values null right, then output a negation
+                if (previousJoinWeight == 0 && rightWeights.joinWeight > 0)
+                {
+                    // Output a negation
+                    weights.Add(-rightWeights.weight);
+                    iterations.Add(0);
+                    for (int i = 0; i < _rightOutputColumns.Count; i++)
+                    {
+                        var outputColumn = columns[_rightOutputIndices[i]];
+                        outputColumn.InsertRangeFrom(outputColumn.Count, rightColumns[_rightOutputColumns[i]], rightIndex, 1);
+                    }
+                    for (int i = 0; i < _leftOutputIndices.Count; i++)
+                    {
+                        columns[_leftOutputIndices[i]].Add(NullValue.Instance);
+                    }
+                }
+                // Went back to be a join with left values and null right
+                if (previousJoinWeight > 0 && rightWeights.joinWeight == 0)
+                {
+                    weights.Add(rightWeights.weight);
+                    iterations.Add(0);
+                    for (int i = 0; i < _rightOutputColumns.Count; i++)
+                    {
+                        var outputColumn = columns[_rightOutputIndices[i]];
+                        outputColumn.InsertRangeFrom(outputColumn.Count, rightColumns[_rightOutputColumns[i]], rightIndex, 1);
+                    }
+                    for (int i = 0; i < _leftOutputIndices.Count; i++)
+                    {
+                        columns[_leftOutputIndices[i]].Add(NullValue.Instance);
+                    }
+                }
+            }
+
+            rightModified = true;
+            leftModified = true;
         }
 
         public override async IAsyncEnumerable<StreamEventBatch> OnRecieve(int targetId, StreamEventBatch msg, long time)
@@ -283,14 +504,16 @@ namespace FlowtideDotNet.Core.Operators.Join.NestedLoopJoin
 
             if (targetId == 0)
             {
-                foreach(var e in msg.Events)
+                for (int i = 0; i < msg.Data.Weights.Count; i++)
                 {
-                    await _leftTemporary.RMW(e, new JoinStorageValue() { Weight = e.Weight, JoinWeight = 0}, (input, current, exist) =>
+                    var rowRef = new ColumnRowReference() { referenceBatch = msg.Data.EventBatchData, RowIndex = i };
+                    await _leftTemporary.RMWNoResult(rowRef, new JoinWeights() { joinWeight = 0, weight = msg.Data.Weights[i] }, (input, current, exist) =>
                     {
                         if (exist)
                         {
-                            current!.Weight += input!.Weight;
-                            if (current.Weight == 0)
+                            current.weight += input.weight;
+
+                            if (current.weight == 0)
                             {
                                 return (default, GenericWriteOperation.Delete);
                             }
@@ -303,14 +526,16 @@ namespace FlowtideDotNet.Core.Operators.Join.NestedLoopJoin
             }
             else if (targetId == 1)
             {
-                foreach (var e in msg.Events)
+                for (int i = 0; i < msg.Data.Weights.Count; i++)
                 {
-                    await _rightTemporary.RMW(e, new JoinStorageValue() { Weight = e.Weight, JoinWeight = 0 }, (input, current, exist) =>
+                    var rowRef = new ColumnRowReference() { referenceBatch = msg.Data.EventBatchData, RowIndex = i };
+                    await _rightTemporary.RMWNoResult(rowRef, new JoinWeights() { joinWeight = 0, weight = msg.Data.Weights[i] }, (input, current, exist) =>
                     {
                         if (exist)
                         {
-                            current!.Weight += input!.Weight;
-                            if (current.Weight == 0)
+                            current.weight += input.weight;
+
+                            if (current.weight == 0)
                             {
                                 return (default, GenericWriteOperation.Delete);
                             }
@@ -330,44 +555,51 @@ namespace FlowtideDotNet.Core.Operators.Join.NestedLoopJoin
         protected override async Task InitializeOrRestore(JoinState? state, IStateManagerClient stateManagerClient)
         {
             Logger.BlockNestedLoopInUse(StreamName, Name);
-            _leftTree = await stateManagerClient.GetOrCreateTree("left", 
-                new BPlusTreeOptions<RowEvent, JoinStorageValue, ListKeyContainer<RowEvent>, ListValueContainer<JoinStorageValue>>()
-            {
-                Comparer = new BPlusTreeListComparer<RowEvent>(new NestedJoinStreamEventComparer()),
-                KeySerializer = new KeyListSerializer<RowEvent>(new StreamEventBPlusTreeSerializer()),
-                ValueSerializer = new ValueListSerializer<JoinStorageValue>(new JoinStorageValueBPlusTreeSerializer()),
-                MemoryAllocator = MemoryAllocator
-            });
-            _rightTree = await stateManagerClient.GetOrCreateTree("right", 
-                new BPlusTreeOptions<RowEvent, JoinStorageValue, ListKeyContainer<RowEvent>, ListValueContainer<JoinStorageValue>>()
-            {
-                Comparer = new BPlusTreeListComparer<RowEvent>(new NestedJoinStreamEventComparer()),
-                KeySerializer = new KeyListSerializer<RowEvent>(new StreamEventBPlusTreeSerializer()),
-                ValueSerializer = new ValueListSerializer<JoinStorageValue>(new JoinStorageValueBPlusTreeSerializer()),
-                MemoryAllocator = MemoryAllocator
-            });
+            _leftTree = await stateManagerClient.GetOrCreateTree("left",
+                new BPlusTreeOptions<ColumnRowReference, JoinWeights, ColumnKeyStorageContainer, JoinWeightsValueContainer>()
+                {
+                    Comparer = new ColumnComparer(_joinRelation.Left.OutputLength),
+                    KeySerializer = new ColumnStoreSerializer(_joinRelation.Left.OutputLength, MemoryAllocator),
+                    ValueSerializer = new JoinWeightsSerializer(MemoryAllocator),
+                    UseByteBasedPageSizes = true,
+                    MemoryAllocator = MemoryAllocator
+                });
+            _rightTree = await stateManagerClient.GetOrCreateTree("right",
+                new BPlusTreeOptions<ColumnRowReference, JoinWeights, ColumnKeyStorageContainer, JoinWeightsValueContainer>()
+                {
+                    Comparer = new ColumnComparer(_joinRelation.Right.OutputLength),
+                    KeySerializer = new ColumnStoreSerializer(_joinRelation.Right.OutputLength, MemoryAllocator),
+                    ValueSerializer = new JoinWeightsSerializer(MemoryAllocator),
+                    UseByteBasedPageSizes = true,
+                    MemoryAllocator = MemoryAllocator
+                });
 
-            _leftTemporary = await stateManagerClient.GetOrCreateTree("left_tmp", 
-                new BPlusTreeOptions<RowEvent, JoinStorageValue, ListKeyContainer<RowEvent>, ListValueContainer<JoinStorageValue>>()
-            {
-                Comparer = new BPlusTreeListComparer<RowEvent>(new NestedJoinStreamEventComparer()),
-                KeySerializer = new KeyListSerializer<RowEvent>(new StreamEventBPlusTreeSerializer()),
-                ValueSerializer = new ValueListSerializer<JoinStorageValue>(new JoinStorageValueBPlusTreeSerializer()),
-                MemoryAllocator = MemoryAllocator
-            });
-
-            _rightTemporary = await stateManagerClient.GetOrCreateTree("right_tmp", 
-                new BPlusTreeOptions<RowEvent, JoinStorageValue, ListKeyContainer<RowEvent>, ListValueContainer<JoinStorageValue>>()
-            {
-                Comparer = new BPlusTreeListComparer<RowEvent>(new NestedJoinStreamEventComparer()),
-                KeySerializer = new KeyListSerializer<RowEvent>(new StreamEventBPlusTreeSerializer()),
-                ValueSerializer = new ValueListSerializer<JoinStorageValue>(new JoinStorageValueBPlusTreeSerializer()),
-                MemoryAllocator = MemoryAllocator
-            });
+            _leftTemporary = await stateManagerClient.GetOrCreateTree("left_tmp",
+                new BPlusTreeOptions<ColumnRowReference, JoinWeights, ColumnKeyStorageContainer, JoinWeightsValueContainer>()
+                {
+                    Comparer = new ColumnComparer(_joinRelation.Left.OutputLength),
+                    KeySerializer = new ColumnStoreSerializer(_joinRelation.Left.OutputLength, MemoryAllocator),
+                    ValueSerializer = new JoinWeightsSerializer(MemoryAllocator),
+                    UseByteBasedPageSizes = true,
+                    MemoryAllocator = MemoryAllocator
+                });
+            _rightTemporary = await stateManagerClient.GetOrCreateTree("right_tmp",
+                new BPlusTreeOptions<ColumnRowReference, JoinWeights, ColumnKeyStorageContainer, JoinWeightsValueContainer>()
+                {
+                    Comparer = new ColumnComparer(_joinRelation.Right.OutputLength),
+                    KeySerializer = new ColumnStoreSerializer(_joinRelation.Right.OutputLength, MemoryAllocator),
+                    ValueSerializer = new JoinWeightsSerializer(MemoryAllocator),
+                    UseByteBasedPageSizes = true,
+                    MemoryAllocator = MemoryAllocator
+                });
 
             if (_eventsProcessed == null)
             {
                 _eventsProcessed = Metrics.CreateCounter<long>("events_processed");
+            }
+            if (_eventsOutCounter == null)
+            {
+                _eventsOutCounter = Metrics.CreateCounter<long>("events");
             }
         }
     }
