+++ tests/FlowtideDotNet.AcceptanceTests/JoinTests.cs
@@ -1148,6 +1148,89 @@ namespace FlowtideDotNet.AcceptanceTests
                 });
         }
 
+        [Fact]
+        public async Task RightJoinMergeJoinFromOrdersUsersFirst()
+        {
+            GenerateCompanies(10);
+            GenerateUsers(1000);
+            //GenerateData(1000);
+            await StartStream(@"
+                INSERT INTO output 
+                SELECT 
+                    o.orderkey, u.firstName, u.LastName
+                FROM orders o
+                RIGHT JOIN users u
+                ON o.userkey = u.userkey");
+            await WaitForUpdate();
+
+            AssertCurrentDataEqual(
+                from user in Users
+                join order in Orders on user.UserKey equals order.UserKey into gj
+                from suborder in gj.DefaultIfEmpty()
+                select new
+                {
+                    suborder?.OrderKey,
+                    user.FirstName,
+                    user.LastName
+                });
+
+            GenerateOrders(1000);
+
+            await WaitForUpdate();
+
+            AssertCurrentDataEqual(
+                from user in Users
+                join order in Orders on user.UserKey equals order.UserKey into gj
+                from suborder in gj.DefaultIfEmpty()
+                select new
+                {
+                    suborder?.OrderKey,
+                    user.FirstName,
+                    user.LastName
+                });
+        }
+
+        [Fact]
+        public async Task RightJoinMergeJoinFromUsersUsersFirst()
+        {
+            GenerateCompanies(10);
+            GenerateUsers(1000);
+            await StartStream(@"
+                INSERT INTO output 
+                SELECT 
+                    o.orderkey, u.firstName, u.LastName
+                FROM users u
+                RIGHT JOIN orders o
+                ON o.userkey = u.userkey");
+            await WaitForUpdate();
+
+            AssertCurrentDataEqual(
+                from order in Orders
+                join user in Users on order.UserKey equals user.UserKey into gj
+                from subuser in gj.DefaultIfEmpty()
+                select new
+                {
+                    order?.OrderKey,
+                    subuser.FirstName,
+                    subuser.LastName
+                });
+
+            GenerateOrders(1000);
+
+            await WaitForUpdate();
+
+            AssertCurrentDataEqual(
+                from order in Orders
+                join user in Users on order.UserKey equals user.UserKey into gj
+                from subuser in gj.DefaultIfEmpty()
+                select new
+                {
+                    order?.OrderKey,
+                    subuser.FirstName,
+                    subuser.LastName
+                });
+        }
+
 
         [Fact]
         public async Task FullOuterJoinMergeJoin()
